pyproject.toml
[tool.poetry]
name = "vaesample"
version = "0.1.0"
description = ""
authors = ["mocobt <mocobt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.19.0"
matplotlib = "^3.3.0"
scipy = "^1.5.1"
tqdm = "^4.48.0"
pyyaml = "^5.3.1"
pytorch-lightning = "^0.8.5"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
